requirements.txt
blue==0.9.1
isort==5.12.0
Pillow==9.5.0
tinydb==4.7.1
ttkbootstrap==1.10.1
